+++ changelog.md
@@ -1,4 +1,7 @@
-## Version 0.7.1 -> 0.7.5
+## Version 0.7.1 -> 0.7.6
+ - Add more special characters to be used for password validation
+ - Add configurable default data path for binds
+ - Remove Redirects from home page
  - Fix compat with non-HTTP protocol like WebDAV (for Nextcloud for example)
  - Fix regression with DNS wildcards certificates
  - Fix issue with the installer when changing both the labels and the volumes